+++ run_tests.py
@@ -3,37 +3,14 @@
 import os
 import sys
 
+parent_dir = os.path.dirname(os.path.abspath(__file__))  # will be: /path/to/django-email-confirm-la/
+sys.path.insert(0, os.path.join(parent_dir, 'email_confirm_la/'))
+sys.path.insert(1, os.path.join(parent_dir, 'test_project/'))
+
+os.environ['DJANGO_SETTINGS_MODULE'] = 'test_project.settings'
+
 from django.conf import settings
 import django
-
-
-parent_dir = os.path.dirname(os.path.abspath(__file__))
-sys.path.insert(0, parent_dir)
-
-settings.configure(
-    DEBUG=False,
-    SECRET_KEY='do_not_go_gentle_into_that_good_night',
-    DATABASES={
-        'default': {
-            'ENGINE': 'django.db.backends.sqlite3',
-            'NAME': ':memory:',
-        },
-    },
-    INSTALLED_APPS=[
-        'django.contrib.auth',
-        'django.contrib.contenttypes',
-        'email_confirm_la',
-    ],
-    MIDDLEWARE_CLASSES=(
-        'django.contrib.sessions.middleware.SessionMiddleware',
-        'django.middleware.common.CommonMiddleware',
-        'django.middleware.csrf.CsrfViewMiddleware',
-        'django.contrib.auth.middleware.AuthenticationMiddleware',
-        'django.contrib.messages.middleware.MessageMiddleware',
-        'django.middleware.clickjacking.XFrameOptionsMiddleware',
-    ),
-    ROOT_URLCONF='',
-)
 
 
 def run_tests():