+++ HISTORY.rst
@@ -1,10 +1,16 @@
 History
 =======
 
+0.2.1 (2014-00-00)
+++++++++++++++++++
+
+- Django 1.4 compatibility: ``update_fields``
+
+
 0.2.0 (2014-11-08)
 ++++++++++++++++++
 
-- Django 1.7 compatibility
+- Django 1.7 compatibility: ``migrations``
 
 
 0.1.0 (2014-10-31)